Moved reader comment service.
